Fix : ResponseDto 리팩토링
