+++ src/main/java/com/example/ondosocial/domain/follower/entity/Follower.java
@@ -1,7 +1,7 @@
 package com.example.ondosocial.domain.follower.entity;
 
 import com.example.ondosocial.config.entity.BaseEntity;
-import com.example.ondosocial.domain.user.entity.User;
+import com.example.ondosocial.domain.profile.entity.User;
 import jakarta.persistence.Entity;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
