+++ src/main/java/com/example/ondosocial/domain/post/entity/Post.java
@@ -1,12 +1,11 @@
 package com.example.ondosocial.domain.post.entity;
 
 import com.example.ondosocial.config.entity.BaseEntity;
-import com.example.ondosocial.domain.user.entity.User;
+import com.example.ondosocial.domain.profile.entity.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.validation.constraints.Min;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
