+++ src/main/java/com/example/ondosocial/domain/profile/dto/response/UserProfileResponseDto.java
@@ -1,0 +1,15 @@
+package com.example.ondosocial.domain.profile.dto.response;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class UserProfileResponseDto {
+    private Long id;
+    private String name;
+    private String email;
+//    private Integer postCount;
+//    private Integer followerCount;
+}
