+++ src/main/java/com/example/ondosocial/domain/profile/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.example.ondosocial.domain.user.repository;
+package com.example.ondosocial.domain.profile.repository;
 
-import com.example.ondosocial.domain.user.entity.User;
+import com.example.ondosocial.domain.profile.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User,Long> {
