--- src/main/java/com/example/ondosocial/domain/user/dto/response/UserProfileDto.java
@@ -1,13 +1,0 @@
-package com.example.ondosocial.domain.user.dto.response;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class UserProfileDto {
-    private Long id;
-    private String name;
-    private String email;
-}
