--- src/main/java/com/example/ondosocial/domain/user/dto/response/UserProfileResponseDto.java
@@ -1,12 +1,0 @@
-package com.example.ondosocial.domain.user.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class UserProfileResponseDto {
-    private String message;
-    private Integer statusCode;
-    private UserProfileDto userProfileDto;
-}
